cleanup and select ResNet
